Added validation and updated.

+++ src/main/java/pl/szlify/codingapi/service/StudentService.java
@@ -2,24 +2,25 @@ package pl.szlify.codingapi.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
+import pl.szlify.codingapi.exceptions.LessonInFutureException;
 import pl.szlify.codingapi.exceptions.MissingStudentException;
 import pl.szlify.codingapi.exceptions.LackofTeacherException;
 import pl.szlify.codingapi.exceptions.BadLanguageException;
-import pl.szlify.codingapi.model.StudentBasicInfoDto;
+import pl.szlify.codingapi.model.*;
+import pl.szlify.codingapi.repository.LessonRepository;
 import pl.szlify.codingapi.repository.StudentRepository;
 import pl.szlify.codingapi.repository.TeacherRepository;
-import pl.szlify.codingapi.model.StudentEntity;
-import pl.szlify.codingapi.model.TeacherEntity;
 import pl.szlify.codingapi.mapper.StudentMapper;
-import pl.szlify.codingapi.model.StudentDto;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
 @RequiredArgsConstructor
 public class StudentService {
     private final StudentRepository studentRepository;
+    private final LessonRepository lessonRepository;
     private final TeacherRepository teacherRepository;
     private final StudentMapper studentMapper;
 
@@ -53,11 +54,12 @@ public class StudentService {
         TeacherEntity teacherEntity = teacherRepository.findByIdAndRemovedFalse(studentBasicInfoDto.getTeacherId())
                 .orElseThrow(LackofTeacherException::new);
 
-        if (!teacherEntity.getLanguages().contains(studentEntity.getLanguage())) {
+        if (!teacherEntity.getLanguages().contains(studentBasicInfoDto.getLanguage())) {
             throw new BadLanguageException();
         }
 
-        StudentEntity updatedStudentEntity = studentMapper.fromBasicInfoAndEntityToEntity(studentEntity, studentBasicInfoDto);
+        StudentEntity updatedStudentEntity = studentMapper
+                .fromBasicInfoAndEntityToEntity(studentEntity, studentBasicInfoDto);
         updatedStudentEntity.setTeacherEntity(teacherEntity);
         return studentMapper.fromEntityToDto(studentRepository.save(updatedStudentEntity));
     }
@@ -65,8 +67,10 @@ public class StudentService {
     public StudentBasicInfoDto updateStudentTeacher(Long id, Long teacherId) {
         StudentEntity studentEntity = studentRepository.findById(id)
                 .orElseThrow(MissingStudentException::new);
+
         TeacherEntity teacherEntity = teacherRepository.findByIdAndRemovedFalse(teacherId)
                 .orElseThrow(LackofTeacherException::new);
+
         if (!teacherEntity.getLanguages().contains(studentEntity.getLanguage())) {
             throw new BadLanguageException();
         }
@@ -76,6 +80,10 @@ public class StudentService {
     }
 
     public void deleteStudent(Long id) {
+        Optional<LessonEntity> existingLesson = lessonRepository.findByStudentEntityId(id);
+        if (existingLesson.isPresent()) {
+            throw new LessonInFutureException();
+        }
         studentRepository.deleteById(id);
     }
 }
